# api/rest/v1/swagger.yaml
definitions:
  github_com_JMURv_effective-mobile_internal_dto.CalculateTotalCostRequest:
    properties:
      from:
        type: string
      service_name:
        maxLength: 255
        minLength: 2
        type: string
      to:
        type: string
      user_id:
        type: string
    required:
    - from
    - service_name
    - to
    - user_id
    type: object
  github_com_JMURv_effective-mobile_internal_dto.CalculateTotalCostResponse:
    properties:
      total_cost:
        maximum: 255
        minimum: 2
        type: integer
    required:
    - total_cost
    type: object
  github_com_JMURv_effective-mobile_internal_dto.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        maxLength: 255
        minLength: 2
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  github_com_JMURv_effective-mobile_internal_dto.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        maxLength: 255
        minLength: 2
        type: string
      start_date:
        type: string
    type: object
  github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse:
    properties:
      errors:
        items:
          type: string
        type: array
    type: object
  github_com_JMURv_effective-mobile_internal_models.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_models.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get a single subscription by UUID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_models.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    post:
      consumes:
      - application/json
      description: Calculate total cost for subscriptions within a period
      parameters:
      - description: Calculation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_dto.CalculateTotalCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_dto.CalculateTotalCostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_JMURv_effective-mobile_internal_hdl_http_utils.ErrorsResponse'
      summary: Calculate total subscription cost
      tags:
      - subscriptions
swagger: "2.0"