# internal/hdl/http/utils/utils.go
package utils

import (
	"errors"
	"fmt"
	"net/http"

	"github.com/JMURv/effective-mobile/internal/ctrl"
	"github.com/JMURv/effective-mobile/internal/hdl"
	"github.com/JMURv/effective-mobile/internal/hdl/validation"
	"github.com/go-playground/validator/v10"
	"github.com/goccy/go-json"
	"go.uber.org/zap"
)

type ErrorsResponse struct {
	Errors []string `json:"errors"`
}

func StatusResponse(w http.ResponseWriter, statusCode int) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(statusCode)
}

func SuccessResponse(w http.ResponseWriter, statusCode int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(statusCode)

	if err := json.NewEncoder(w).Encode(data); err != nil {
		zap.L().Error("failed to encode success response", zap.Error(err))
		return
	}
}

func ErrResponse(w http.ResponseWriter, statusCode int, err error) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(statusCode)

	msgs := make([]string, 0, 1)
	if errs, ok := err.(validator.ValidationErrors); ok {
		msgs = make([]string, 0, len(errs))
		for _, fe := range errs {
			msgs = append(msgs, fmt.Sprintf("%s failed on the %s rule", fe.Field(), fe.Tag()))
		}
	} else {
		msgs = append(msgs, err.Error())
	}

	if err := json.NewEncoder(w).Encode(
		&ErrorsResponse{
			Errors: msgs,
		},
	); err != nil {
		zap.L().Error("failed to encode err response", zap.Error(err))
		return
	}
}

func ParseAndValidate(w http.ResponseWriter, r *http.Request, dst any) bool {
	var err error

	if err = json.NewDecoder(r.Body).Decode(dst); err != nil {
		zap.L().Error(
			hdl.ErrDecodeRequest.Error(),
			zap.Error(err),
		)
		ErrResponse(w, http.StatusBadRequest, hdl.ErrDecodeRequest)

		return false
	}

	if err = validation.V.Struct(dst); err != nil {
		zap.L().Debug("failed to validate request", zap.Error(err))
		ErrResponse(w, http.StatusBadRequest, err)
		return false
	}

	return true
}

func HandleError(w http.ResponseWriter, err error) {
	if errors.Is(err, ctrl.ErrNotFound) {
		ErrResponse(w, http.StatusNotFound, err)
		return
	}

	if errors.Is(err, ctrl.ErrAlreadyExists) {
		ErrResponse(w, http.StatusConflict, err)
		return
	}

	ErrResponse(w, http.StatusInternalServerError, hdl.ErrInternal)
	return
}